% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{aiEst}
\alias{aiEst}
\title{Estimate adverse impact given d and sr}
\usage{
aiEst(d, sr, pct_minority)
}
\arguments{
\item{d}{Subgroup difference.}

\item{sr}{The percentage of the applicant population that is selected.}

\item{pct_minority}{The percentage of the applicant population that is part of
a given minority group.}
}
\value{
\describe{
    \item{ai}{The adverse impact ratio.}
    \item{overall_sr}{The overall selection ratio.}
    \item{sr_majority}{The selection ratio for the majority group.}
    \item{sr_minority}{The selection ratio for the minority group.}
    \item{uc}{the predictor cutoff value that corresponds to the given overall selection ratio}
  }
}
\description{
Estimate adverse impact given d and sr
}
\examples{
aiEst(d = -0.15, sr = 0.25, pct_minority = 0.30)

aiEst(d = -0.40, sr = 0.10, pct_minority = 0.15)
}
\author{
Jeff Jones and Allen Goebl
}
\references{
De Corte, W., Lievens, F.(2003). A practical procedure to estimate
the quality and the adverse impact of single-stage selection decisions.
\emph{International Journal of Selection and Assessment, 11(1)}, 87-95.
}

