% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asvab}
\alias{asvab}
\title{Wee, Newman, & Joseph, (2014) ASVAB data}
\format{A data frame with 12 rows and 4 columns.}
\usage{
asvab
}
\description{
This dataset was published in Wee, Newman, and Joseph (2014)
and describes the results of a military validation study. The first four
rows contain the intercorrelations of the four predictor variables. The fifth
row contains the black-white score differences (d). Rows 6-12 contain the 
correlations between the four predictor variables and the six job performance
variables.
}
\references{
Wee, S., Newman, D. A., & Joseph, D. L. (2014). More than g: 
Selection quality and adverse impact implications of considering second-stratum
cognitive abilities. \emph{Journal of Applied Psychology, 99(4)}.
}
\keyword{datasets}

