% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{cor2d}
\alias{cor2d}
\title{Convert from r to d}
\usage{
cor2d(r, n_1 = 1, n_2 = 1)
}
\arguments{
\item{r}{A r-value or a vector of r values.}

\item{n_1}{The sample size of group 1.}

\item{n_2}{The sample size of group 2.}
}
\value{
A d value or a vector of d values.
}
\description{
Convert from r to d
}
\examples{
cor2d(.3)
cor2d(.3, n_1 = 20, n_2 = 50)
cor2d(((1:9)/10))
}
\author{
Jeff Dahlke, Allen Goebl, and Jeff Jones
}
