% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dls2007}
\alias{dls2007}
\title{Decorte, Lievens, & Sackett (2007) example data}
\format{A data frame with 6 rows and 7 columns.}
\usage{
dls2007
}
\description{
This hypothetical dataset was published in Decorte, W., Lievens, F., Sackett,
P. R. (2007). The first column contains black-white subgroup difference scores.
Columns 2-7 contain a hypothetical predictor, job performance correlation matrix.
}
\references{
De Corte, W., Lievens, F., & Sackett, P. R. (2007) Combining
predictors to achieve optimal trade-offs between selection quality and adverse
impact. emph{Journal of Applied Psychology}, 92(5), 1380.
}
\keyword{datasets}

