% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse2}
\alias{fuse2}
\title{Computes the correlation between two composites of items using weights.}
\usage{
fuse2(r_mat, wt_a, wt_b)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{wt_a}{A vector containing the weights of each item in composite A. 
Items which are not included in the composite should be assigned a
weight of 0.}

\item{wt_b}{A vector containing the weights of each item in composite B. 
Items which are not included in the composite should be assigned a
weight of 0.}
}
\value{
A correlation coefficient.
}
\description{
Computes the correlation between two composites of items. Composites may 
contain overalapping items. Items weights for each composite may be specified.
}
\note{
This is an alternative version of \code{fuse()} which uses weight vectors
      to specify both item selection and weight. This syntax maybe be preferable
      to some users. Furthermore, this function is more powerful in that it
      can return values for multiple sets of weights.
}
\examples{
Rxx <- matrix(c(1.00, 0.25,  0.50,  0.61,
                0.25, 1.00,  0.30,  0.10,
                0.50, 0.30,  1.00, -0.30,
                0.61, 0.10, -0.30,  1.00), 4, 4)
wt_a   <- c(1, 0, 1, 0)
wt_b   <- c(0, 1, 0, 1)

# Example using overlapping items and weights
Rxx  <- matrix(.3, 4, 4); diag(Rxx) <- 1
wt_a <- c(.60, .25, 0, .15)
wt_b <- c(0, 2, 3, 0)

fuse2(r_mat = Rxx, wt_a = wt_a, wt_b = wt_b)

}
\references{
Lord, F.M. & Novick, M.R. (1968). \emph{Statisticl theories of 
            menal test scores}.
}
\author{
Allen Goebl and Jeff Jones
}
