% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.checkIndex}
\alias{.checkIndex}
\title{Are y_col and x_col appropriate indexs for r_mat?}
\usage{
.checkIndex(r_mat, y_col, x_col)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{y_col}{A vector of columns representing criterion variables.}

\item{x_col}{A vector of columns representing predictor variables.}
}
\value{
Nothing. Will result in an error if test fails.
}
\description{
Are y_col and x_col appropriate indexs for r_mat?
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

