% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.corAdd}
\alias{.corAdd}
\title{Appends a new variable into a correlation matrix.}
\usage{
.corAdd(r_mat, r_vec, lab = "")
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{r_vec}{A vector of correlations to append to r_mat.}

\item{lab}{A column name for r_vec.}
}
\value{
A larger correlation matrix.
}
\description{
Appends a new variable into a correlation matrix.
}
\examples{
#data(dls2007)
#dat <- dls2007
#rxx <- dat[1:4, 2:5]
#corAdd(rxx, c(.1,.1,.1,.1), lab="V5")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
