% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{.fuseCom}
\alias{.fuseCom}
\title{Computes the intercorrelations of item composites}
\usage{
.fuseCom(r_mat, key_mat)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{key_mat}{A matrix with one row for each composite and one column for 
each item contained in r_mat. The value if each element corresponds
to the weight given to an item.}
}
\value{
A matrix of intercorrelations.
}
\description{
The key matrix is used to specify any number of weighted item composites.
A correlation matrix of these composites is then computed and returned.
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

