% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{trUtility}
\alias{trUtility}
\title{Taylor-Russell Utility}
\usage{
trUtility(n = 1, rxy, sr, br, dbr, cost = 0, period = 1)
}
\arguments{
\item{n}{The size of the applicant pool.}

\item{rxy}{The correaltion between the predictor composite and the criterion.}

\item{sr}{The selection ratio.}

\item{br}{The criterion ratio. The cutoff point indicating success or failure.}

\item{dbr}{The monetary value of a 1 percent change in the basis rate per applicant.}

\item{cost}{The cost per applicant of a selection system.}

\item{period}{The anticipated tenure of selected employees}
}
\value{
Estimated gain in utility.
}
\description{
The Taylor Russel Model can be used to estimate the utility of selecting
for a dichotomous criterion.
}
\examples{
#trUtility(rxy=.5, sr=.5, br=.6, dbr=1000)
}
\author{
Allen Goebl and Jeff Jones
}
\references{
Roomsburg (1989). \emph{Utility as a function of selection ratio
and base rate: An empirical investigation of military aviation selection.}
(Doctoral dissertation).
}
\keyword{internal}

