% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.indexMat}
\alias{.indexMat}
\title{Finds rxx and rxy for a correlation matrix}
\usage{
.indexMat(r_mat, y_col, x_col)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{y_col}{A vector of columns representing criterion variables.}

\item{x_col}{A vector of columns representing predictor variables.}
}
\value{
Matrix rxx, and vector rxy.
}
\description{
Finds rxx and rxy for a correlation matrix
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

