% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.isCorMat}
\alias{.isCorMat}
\title{Is a matrix a correlation matrix?}
\usage{
.isCorMat(r_mat)
}
\arguments{
\item{r_mat}{A correlation matrix.}
}
\value{
Nothing. Will return an error if \code{r_mat} is not a correlation matrix
}
\description{
Is a matrix a correlation matrix?
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
