% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{.paretoXX}
\alias{.paretoXX}
\title{Computes data needed for a XX Pareto plot.}
\usage{
.paretoXX(r_mat, x_col, y_col, pts = 100)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{x_col}{A vector of columns representing predictor variables.}

\item{y_col}{A vector of columns representing criterion variables.}

\item{pts}{The number of points used. Determines accuracy.}
}
\value{
Points along the pareto optimal surface and the predictor weights
        used to compute them.
}
\description{
Computes data needed for a XX Pareto plot.
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

