% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_importance.R
\name{.relWt}
\alias{.relWt}
\title{Relative weights}
\usage{
.relWt(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor, criterion correlations.}
}
\value{
DO THIS JEFF
}
\description{
Function to implement Johnson's (2000) relative weight computation.
}
\examples{
print("example needed")
}
\author{
Jeff Jones and Allen Goebl
}
\references{
Johnson, J. (2000). A heuristic method for estimating the
            relative weight of predictor variables in multiple regression.
            \emph{Multivariate Behavioral Research, 35}, 1--19.
}
\keyword{internal}

