% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{.rmatBeta}
\alias{.rmatBeta}
\title{Find regression weights}
\usage{
.rmatBeta(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor, criterion correlations.}
}
\value{
A vector of regression weights.
}
\description{
Find regression weights
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

