% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{.rmatBetaPE}
\alias{.rmatBetaPE}
\title{Partially evaluated regression}
\usage{
.rmatBetaPE(rxx)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}
}
\value{
Partially evaluated regression function.
}
\description{
Returns a function for calculating beta weights which has been partially
evalauted with respect to rxx.
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

