% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{.rmatReg}
\alias{.rmatReg}
\title{Find regression weights and R2}
\usage{
.rmatReg(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor, criterion correlations.}
}
\value{
R2 and Regression weights
}
\description{
Find regression weights and R2
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}

