% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.unpack}
\alias{.unpack}
\title{Unpacks key vector}
\usage{
.unpack(key_vec)
}
\arguments{
\item{key_vec}{A key vector.}
}
\value{
Returns (1) A list of indices (2) A list of standardized weights.
}
\description{
Unpacks key vector
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
