% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.unpackMat}
\alias{.unpackMat}
\title{Unpacks key matrix}
\usage{
.unpackMat(key_mat)
}
\arguments{
\item{key_mat}{A matrix of keys. Each row is a key.}
}
\value{
Returns (1) A list of indices (2) A list of standardized weights.
}
\description{
Works like .unpack but accepts matrices rather than vectors.
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
