% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.utilitySwitch}
\alias{.utilitySwitch}
\title{Utility Input Switch}
\usage{
.utilitySwitch(pux = NULL, uxs = NULL, rxy = NULL, sr = NULL)
}
\arguments{
\item{pux}{The expected average criterion score of selected applicants.}

\item{uxs}{The average predicter score of those selected}

\item{rxy}{The correlation between the predictor composite and the criterion.}

\item{sr}{A selection ratio or a vector of selection ratios.}
}
\value{
The expected average criterion score of selected applicants.
}
\description{
Utility functions
}
\note{
Many utility functions can except either (1) pux, (2) uxs and rxy, or
      (3) sr and rxy.
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
