% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{.wtPair}
\alias{.wtPair}
\title{An internal function for generating criterion weights for XX pareto plots.}
\usage{
.wtPair(pts)
}
\arguments{
\item{pts}{How many different pairs of criterion weights to calculate.}
}
\value{
A matrix of of criterion weights with one column per criteiron.
}
\description{
An internal function for generating criterion weights for XX pareto plots.
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
\keyword{internal}
