% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcea1994}
\alias{rcea1994}
\title{Ree, Carretta, Earles, Albert (1994)}
\format{A list containing a 4x4 matrix and a 2x2 matrix as elements.}
\usage{
rcea1994
}
\description{
This example data was published in Ree, M. J., Carretta, T. R., Earles, J. A.,
& Albert, W. (1994). The data set contains two matrices stored as a list,
which can be used to demonstrate multivariate range restriction corrections. The
vstar matrix is the variance-covariance matrix of the unrestricted sample. The 
vpp matrix is the variance covariance matrix of the restricted sample. The 
vpp matrix represents the subset of variables which were explicitly used for
selection, which are also found in the upper left corner of the vstar matrix.
}
\references{
Ree, M. J., Carretta, T. R., Earles, J. A., & Albert, W. (1994).
Sign changes when correcting for range restriction: A note on Pearson's and
Lawley's selection formulas. \emph{Journal of Applied Psychology}, 72(2), 298.
}
\keyword{datasets}

