% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliabate}
\alias{reliabate}
\title{Disattenuate a correlation matrix using an estimate of the component reliabilities}
\usage{
reliabate(r_mat, rel_vec)
}
\arguments{
\item{r_mat}{A correlation matrix}

\item{rel_vec}{A vector or reliabilities.}
}
\value{
A reliabated (disattenauted) correlation matrix.
}
\description{
Disattenuate a correlation matrix using an estimate of the component reliabilities
}
\examples{
r_mat <- matrix(c(1.00, 0.25, 0.30, 
                  0.25, 1.00, 0.50, 
                  0.30, 0.50, 1.00), 3, 3)
rel   <- c(.70, .64, .81)
reliabate(r_mat = r_mat, rel_vec = rel)
}
\author{
Allen Goebl and Jeff Jones
}
