% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{rmatReg}
\alias{rmatReg}
\title{Regression}
\usage{
rmatReg(r_mat, y_col, x_col)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{y_col}{The column representing the criterion variable.}

\item{x_col}{A vector of columns representing predictor variables.}
}
\value{
Regression beta weights and R2.
}
\description{
Regression
}
\examples{
Rs <- matrix(c(1.0, 0.2,  0.3, 0.4, -0.4,
              0.2, 1.0,  0.5, 0.1,  0.1,
              0.3, 0.5,  1.0, 0.2, -0.3,
              0.4, 0.1,  0.2, 1.0,  0.4,
             -0.4, 0.1, -0.3, 0.4,  1.0), 5, 5)
ys <- 5
xs <- 1:4

rmatReg(Rs, ys, xs)
}
\author{
Allen Goebl and Jeff Jones
}

