% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{rmatRegPE}
\alias{rmatRegPE}
\title{Partially evaluated regression}
\usage{
rmatRegPE(rxx)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}
}
\value{
Partially evaluated regression function.
}
\description{
Returns a function for calculating beta weights and R2 which has been 
partially evalauted with respect to rxx.
}
\examples{
Rxx <- matrix(c(1.00, 0.25, 0.40,
                0.25, 1.00, 0.30,
                0.40, 0.30, 1.00), 3, 3)

rmatRegPE(Rxx)
}
\author{
Allen Goebl and Jeff Jones
}

