% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{solveBeta}
\alias{solveBeta}
\title{Find beta weights}
\usage{
solveBeta(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor criterion intercorrelations, or a
matrix with one row per criterion.}
}
\value{
A vector or matrix of regression weights.
}
\description{
Find beta weights
}
\examples{
library(iopsych)
data(dls2007)
dat <- dls2007[1:6, 2:7]
rxx <- dat[1:4, 1:4]
rxy <- dat[1:4, 5]

solveBeta(rxx=rxx, rxy=rxy)
}
\author{
Allen Goebl and Jeff Jones
}
