% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{solveR2}
\alias{solveR2}
\title{Find R2}
\usage{
solveR2(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{rxy A vector of predictor criterion intercorrelations, or a
matrix with one row per criterion.}
}
\value{
R2 and Regression weights
}
\description{
Find R2
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}

