% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{solveReg}
\alias{solveReg}
\title{Find beta weights and R2}
\usage{
solveReg(rxx, rxy)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor criterion intercorrelations, or a
matrix with one row per criterion.}
}
\value{
R2 and Regression weights
}
\description{
Find beta weights and R2
}
\examples{
print("example needed")
}
\author{
Allen Goebl and Jeff Jones
}
