% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{solveWtPred}
\alias{solveWtPred}
\title{Correlation between weighted predictor composite and criterion.}
\usage{
solveWtPred(rxx, rxy, wt)
}
\arguments{
\item{rxx}{A matrix of predictor intercorrelations.}

\item{rxy}{A vector of predictor criterion intercorrelations, or a
matrix with one row per criterion.}

\item{wt}{A vector of predictor weights, or a matrix with one set of 
predictor weights per row.}
}
\value{
A matrix of correlation coefficent with one row per weight vector
        and one column per rxy vector.
}
\description{
Correlation between weighted predictor composite and criterion.
}
\examples{
library(iopsych)
data(dls2007)
dat <- dls2007[1:6, 2:7]
rxx <- dat[1:4, 1:4]
rxy <- dat[5:6, 1:4]

wt1 <- c(1,1,1,1)
wt2 <- c(1,2,3,4)
wt_mat <- rbind(wt1, wt2)

solveWtPred(rxx=rxx, rxy=rxy, wt=wt_mat)
}
\author{
Allen Goebl Jeff Jones
}
