% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{trModel}
\alias{trModel}
\title{Taylor-Russell Ratio}
\usage{
trModel(rxy, sr, br)
}
\arguments{
\item{rxy}{The correaltion between the predictor composite and the criterion.}

\item{sr}{The selection ratio.}

\item{br}{The base rate of the criterion. The cutoff point indicating
success or failure.}
}
\value{
The success ratio.
}
\description{
Computes the Taylor Russel ratio
}
\examples{
trModel(rxy=.5, sr=.5, br=.6)
}
\author{
Allen Goebl and Jeff Jones
}
\references{
Taylor, H. C., & Russell, J. T. (1939). The relationship of
validity coefficients to the practical effectiveness of tests in selection:
Discussion and tables. \emph{Journal of Applied Psychology}, 25(5), 565.
}

