% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{ux}
\alias{ux}
\title{The average score of selected applicants on a predictor composite.}
\usage{
ux(sr)
}
\arguments{
\item{sr}{A selection ratio or a vector of selection ratios.}
}
\value{
ux: The average score of those selected on a predicter composite.
}
\description{
When scores on the predictor composite are assumed to be normally
distributed, the average score of selected applicants can be computed for 
an arbitrary selection ratio using the ordinate of the normal curve.
}
\examples{
ux(.6)
}
\author{
Allen Goebl and Jeff Jones
}
\references{
Naylor, J. C., & Shine, L. C. (1965). A table for determining the
 increase in mean criterion score obtained by using a selection device. 
 \emph{Journal of Industrial Psychology}, 78-109.
}

