% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{cor2d}
\alias{cor2d}
\title{Convert from r to d}
\usage{
cor2d(r)
}
\arguments{
\item{r}{A r-value or a vector of r values.}
}
\value{
A d value or a vector of d values.
}
\description{
Convert from r to d
}
\examples{
cor2d(.3)
cor2d(((1:9)/10))
}
\author{
Allen Goebl and Jeff Jones
}

