% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{d2cor}
\alias{d2cor}
\title{Convert from d to r}
\usage{
d2cor(d)
}
\arguments{
\item{d}{A d-value or a vector of d values.}
}
\value{
A r value or a vector of r values.
}
\description{
Convert from d to r
}
\examples{
d2cor(.3)
d2cor(((1:9)))
}
\author{
Allen Goebl and Jeff Jones
}

