% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuseVec}
\alias{fuseVec}
\title{Computes the correlations between a correlation matrix and a weighted composite
of items from the matrix.}
\usage{
fuseVec(r_mat, wt)
}
\arguments{
\item{r_mat}{A correlation matrix.}

\item{wt}{A vector containing the weights of each item in composite A or a
matrix with one row per weight vector.}
}
\value{
A vector of correlation coefficients will be returned if \code{wt_a}
        is a vector. If /code{wt_b} is a matrix, a matrix of correlation 
        coefficients with one row for each weight vector will be returned.
}
\description{
Computes the correlations between a correlation matrix and a weighted composite
of items from the matrix.
}
\examples{
data(dls2007)
dat <- dls2007
rxx <- dat[1:4, 2:5]
wt1 <- c(1, 1, 1, 1)
wt2 <- c(2, 0, 1, 0)
wt  <- rbind(wt1, wt2)

fuseVec(r_mat=rxx, wt=wt1)
}
\author{
Allen Goebl and Jeff Jones
}
\references{
Lord, F.M. & Novick, M.R. (1968). \emph{Statisticl theories of 
            mental test scores}.
}

